Rename deployments modules
