+++ gaphor/UML/deployments/deploymentpropertypage.py
